Refactor
- Refactor: Moved the responses tests to test_basic module (to be renamed)
- Refactor: Moved the requests-mock tests to test_advanced module (to be renamed)
- Bug Fix: Renamed classes for each test to prevent conflicts

Note: This fixes the Py2 and Pypy tests that were broken
