+++ stackinabox/tests/test_basic.py
@@ -1,19 +1,18 @@
 """
 Stack-In-A-Box: Basic Test
 """
+import json
 import logging
 import unittest
 
-import httpretty
 import requests
 import responses
 import six
 
-import stackinabox.util_httpretty
 import stackinabox.util_responses
-import stackinabox.util_requests_mock
 from stackinabox.stack import StackInABox
 from stackinabox.services.hello import HelloService
+from stackinabox.tests.utils.services import AdvancedService
 
 
 logger = logging.getLogger(__name__)
@@ -35,32 +34,36 @@ def test_basic_responses():
     run()
 
 
-class TestRequestsMock(unittest.TestCase):
+@unittest.skipIf(six.PY3, 'Responses fails on PY3')
+def test_advanced_responses():
 
-    def setUp(self):
-        super(TestRequestsMock, self).setUp()
-        StackInABox.register_service(HelloService())
-        self.session = requests.Session()
+    def run():
+        responses.mock.start()
+        StackInABox.register_service(AdvancedService())
+        stackinabox.util_responses.responses_registration('localhost')
 
-    def tearDown(self):
-        super(TestRequestsMock, self).tearDown()
+        res = requests.get('http://localhost/advanced/')
+        assert res.status_code == 200
+        assert res.text == 'Hello'
+
+        res = requests.get('http://localhost/advanced/h')
+        assert res.status_code == 200
+        assert res.text == 'Good-Bye'
+
+        expected_result = {
+            'bob': 'bob: Good-Bye alice',
+            'alice': 'alice: Good-Bye bob',
+            'joe': 'joe: Good-Bye jane'
+        }
+        res = requests.get('http://localhost/advanced/g?bob=alice;'
+                           'alice=bob&joe=jane')
+        assert res.status_code == 200
+        assert res.json() == expected_result
+
         StackInABox.reset_services()
-        self.session.close()
 
-    def test_basic_requests_mock(self):
-        stackinabox.util_requests_mock.requests_mock_session_registration(
-            'localhost', self.session)
+        responses.mock.stop()
+        responses.mock.reset()
 
-        res = self.session.get('http://localhost/hello/')
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(res.text, 'Hello')
-
-    def test_context_requests_mock(self):
-        with stackinabox.util_requests_mock.activate():
-
-            stackinabox.util_requests_mock.requests_mock_registration(
-                'localhost')
-
-            res = requests.get('http://localhost/hello/')
-            self.assertEqual(res.status_code, 200)
-            self.assertEqual(res.text, 'Hello')
+    with responses.RequestsMock():
+        run()
